Version control all server logic
