+++ server/__init__.py
@@ -1,7 +1,1 @@
-from .apps import *
-from .config import *
-from .database import *
-from .errors import *
-from .models import *
-from .routes import *
-from .utils import *
+from .v1 import *
