--- server/errors.py
@@ -1,11 +1,0 @@
-from __future__ import annotations
-
-
-__all__ = (
-    "ResidentManagerException",
-)
-
-
-class ResidentManagerException(Exception):
-    """Base class for all exceptions from this server"""
-    pass
