--- server/routes/api/__init__.py
@@ -1,1 +1,0 @@
-from .v1 import *
