+++ server/v1/__init__.py
@@ -1,0 +1,6 @@
+from .apps import *
+from .config import *
+from .database import *
+from .models import *
+from .routes import *
+from .utils import *
