+++ server/v1/models/residents.py
