+++ server/v1/models/rooms.py
