+++ server/v1/routes/api/admin/login.py
@@ -1,25 +1,25 @@
 from __future__ import annotations
 
-from typing import List, Optional
+from typing import Optional
 
 from fastapi import Response, status
 
-from ......apps import api_v1
-from ......models import AuthorizationHeader, Result, RoomData
+from ....apps import api_v1
+from ....models import AuthorizationHeader, Result
 
 
-__all__ = ("admin_rooms_delete",)
+__all__ = ("admin_login",)
 
 
 @api_v1.post(
-    "/admin/rooms/delete",
-    name="Room information deletion",
-    description="Delete room information",
+    "/admin/login",
+    name="Administrators login",
+    description="Verify administrator authorization data.",
     tags=["admin"],
     response_model=None,
     responses={
         status.HTTP_204_NO_CONTENT: {
-            "description": "Operation completed successfully",
+            "description": "Successfully logged in",
         },
         status.HTTP_400_BAD_REQUEST: {
             "description": "Incorrect authorization data",
@@ -28,15 +28,12 @@ __all__ = ("admin_rooms_delete",)
     },
     status_code=status.HTTP_204_NO_CONTENT,
 )
-async def admin_rooms_delete(
+async def admin_login(
     headers: AuthorizationHeader,
     response: Response,
-    rooms: List[int],
 ) -> Optional[Result[None]]:
-    auth = await headers.verify_admin()
-    if auth is not None:
+    result = await headers.verify_admin()
+    if result is not None:
         response.status_code = status.HTTP_400_BAD_REQUEST
-        return auth
 
-    await RoomData.delete_many(rooms)
-    return None
+    return result
