+++ server/v1/routes/api/admin/residents/__init__.py
