+++ server/v1/routes/api/admin/residents/delete.py
@@ -1,25 +1,25 @@
 from __future__ import annotations
 
-from typing import Optional
+from typing import List, Optional
 
 from fastapi import Response, status
 
 from .....apps import api_v1
-from .....models import AuthorizationHeader, Result
+from .....models import AuthorizationHeader, Resident, Result, Snowflake
 
 
-__all__ = ("admin_login",)
+__all__ = ("admin_residents_delete",)
 
 
 @api_v1.post(
-    "/admin/login",
-    name="Administrators login",
-    description="Verify administrator authorization data.",
+    "/admin/residents/delete",
+    name="Account deletion",
+    description="Delete one or more resident accounts",
     tags=["admin"],
     response_model=None,
     responses={
         status.HTTP_204_NO_CONTENT: {
-            "description": "Successfully logged in",
+            "description": "Operation completed successfully",
         },
         status.HTTP_400_BAD_REQUEST: {
             "description": "Incorrect authorization data",
@@ -28,12 +28,15 @@ __all__ = ("admin_login",)
     },
     status_code=status.HTTP_204_NO_CONTENT,
 )
-async def admin_login(
+async def admin_residents_delete(
     headers: AuthorizationHeader,
     response: Response,
+    objects: List[Snowflake],
 ) -> Optional[Result[None]]:
-    result = await headers.verify_admin()
-    if result is not None:
+    auth = await headers.verify_admin()
+    if auth is not None:
         response.status_code = status.HTTP_400_BAD_REQUEST
+        return auth
 
-    return result
+    await Resident.delete_many(objects)
+    return None
