+++ server/v1/routes/api/admin/residents/root.py
@@ -4,9 +4,9 @@ from typing import List, Literal, Optional
 
 from fastapi import Response, status
 
-from ......apps import api_v1
-from ......config import DB_PAGINATION_QUERY
-from ......models import AuthorizationHeader, Resident, Result
+from .....apps import api_v1
+from .....config import DB_PAGINATION_QUERY
+from .....models import AuthorizationHeader, Resident, Result
 
 
 __all__ = ("admin_residents",)
