+++ server/v1/routes/api/admin/rooms/root.py
@@ -4,9 +4,9 @@ from typing import List, Optional
 
 from fastapi import Response, status
 
-from ......apps import api_v1
-from ......config import DB_PAGINATION_QUERY
-from ......models import AuthorizationHeader, Result, Room
+from .....apps import api_v1
+from .....config import DB_PAGINATION_QUERY
+from .....models import AuthorizationHeader, Result, Room
 
 
 __all__ = ("admin_rooms",)
