+++ server/v1/routes/api/login.py
@@ -4,8 +4,8 @@ from typing import Optional, cast
 
 from fastapi import Response, status
 
-from ....apps import api_v1
-from ....models import AuthorizationHeader, PublicInfo, Resident, Result
+from ...apps import api_v1
+from ...models import AuthorizationHeader, PublicInfo, Resident, Result
 
 
 __all__ = ("login",)
