+++ server/v1/routes/api/residents/__init__.py
